move all route components
