--- ui/src/app/Layout.tsx
@@ -1,38 +1,0 @@
-import { Box, Flex, Heading } from "@chakra-ui/react";
-import { useEffect, useState } from "react";
-import { Link, Outlet } from "react-router-dom";
-import Sidebar from "../components/Sidebar";
-import isLoggedIn from "../hooks/isLoggedIn";
-
-export default function Home() {
-  const [loggedIn, setIsLoggedIn] = useState<boolean>(false);
-  useEffect(() => {
-    setIsLoggedIn(isLoggedIn())
-  }, [loggedIn])
-
-  const authenticatedLayout = (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24" >
-      <Box width={'100%'} bg={'gray.50'} borderBottom={'1px solid #f3f3f3'} padding={4}>
-        <Link to="/">
-          <Heading size='1xl'>QATARINA </Heading>
-        </Link>
-      </Box>
-      <Flex>
-        <Box w="20%" borderRight={'1px solid rgba(200, 200, 200, 1)'} h="100vh">
-          <Sidebar />
-        </Box>
-        <Box padding={4}>
-          <Outlet />
-        </Box>
-      </Flex>
-    </main>
-  );
-
-  return (
-    <>
-      {loggedIn ?
-        (authenticatedLayout) :
-        (<Outlet />)}
-    </>
-  );
-}
