--- ui/src/app/dashboard/index.tsx
@@ -1,9 +1,0 @@
-import { IconDashboard } from "@tabler/icons-react";
-
-export default function DashboardPage() {
-  return (
-    <div>
-      <h1><IconDashboard /> Dashboard</h1>
-    </div>
-  )
-}
