--- ui/src/app/integrations/ListIntegrations.tsx
@@ -1,9 +1,0 @@
-import { Box, Heading } from "@chakra-ui/react";
-
-export default function ListIntegrations() {
-  return (
-    <Box>
-      <Heading>Integrations</Heading>
-    </Box>
-  )
-}
