--- ui/src/app/projects/ExecuteTestPlan.tsx
@@ -1,9 +1,0 @@
-import TestCaseGrid from '../../components/TestCaseGrid'
-
-export default function ExecuteTestPlan() {
-    return (
-        <div>
-            <TestCaseGrid />
-        </div>
-    )
-}
