--- ui/src/app/projects/[id]/testcases/view/[testCaseId]/page.tsx
@@ -1,52 +1,0 @@
-'use client';
-
-import { Tab, TabList, TabPanel, TabPanels, Tabs } from '@chakra-ui/react';
-
-export default function ViewTestCase() {
-  return (
-    <div className="card">
-      <Tabs>
-        <TabList>
-          <Tab>Description</Tab>
-          <Tab>Testers</Tab>
-          <Tab>Test Results</Tab>
-          <Tab>Documents and Media</Tab>
-        </TabList>
-        <TabPanels>
-          <TabPanel title="Description">
-            <p className="m-0">
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
-              Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
-              consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-            </p>
-          </TabPanel>
-          <TabPanel title="Testers">
-            <p className="m-0">
-              Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam,
-              eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo
-              enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui
-              ratione voluptatem sequi nesciunt. Consectetur, adipisci velit, sed quia non numquam eius modi.
-            </p>
-          </TabPanel>
-          <TabPanel title="Test Results">
-            <p className="m-0">
-              Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam,
-              eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo
-              enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui
-              ratione voluptatem sequi nesciunt. Consectetur, adipisci velit, sed quia non numquam eius modi.
-            </p>
-          </TabPanel>
-          <TabPanel title="Documents and Media">
-            <p className="m-0">
-              At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis praesentium voluptatum deleniti atque corrupti
-              quos dolores et quas molestias excepturi sint occaecati cupiditate non provident, similique sunt in
-              culpa qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio.
-              Nam libero tempore, cum soluta nobis est eligendi optio cumque nihil impedit quo minus.
-            </p>
-          </TabPanel>
-        </TabPanels>
-      </Tabs>
-    </div>
-  )
-}
