--- ui/src/app/reports/ListReports.tsx
@@ -1,9 +1,0 @@
-import { Box, Heading } from "@chakra-ui/react";
-
-export default function ListReports() {
-  return (
-    <Box>
-      <Heading>Reports</Heading>
-    </Box>
-  )
-}
