--- ui/src/app/settings/page.tsx
@@ -1,7 +1,0 @@
-export default function SettingsIndex() {
-  return (
-    <div>
-      <h1>Settings</h1>
-    </div>
-  )
-}
