--- ui/src/app/testers/ListTesters.tsx
@@ -1,12 +1,0 @@
-import { Box, Button, Heading } from "@chakra-ui/react";
-
-export default function ListTesters() {
-    return (
-        <Box>
-            <Heading>Testers</Heading>
-
-            <Button>Add New Tester</Button>
-            <Button>Invite a Tester</Button>
-        </Box>
-    )
-}
