+++ ui/src/lib/providers.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import { ChakraProvider } from '@chakra-ui/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
