+++ ui/src/routes/(app)/testers/index.tsx
@@ -1,9 +1,17 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { Box, Button, Heading } from "@chakra-ui/react";
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute('/(app)/testers/')({
-  component: RouteComponent,
-})
+export const Route = createFileRoute("/(app)/testers/")({
+  component: ListTesters,
+});
 
-function RouteComponent() {
-  return <div>Hello "/testers/"!</div>
+function ListTesters() {
+  return (
+    <Box>
+      <Heading>Testers</Heading>
+
+      <Button>Add New Tester</Button>
+      <Button>Invite a Tester</Button>
+    </Box>
+  );
 }
