+++ ui/src/routes/(auth)/logout.tsx
@@ -1,9 +1,19 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute, useNavigate } from "@tanstack/react-router";
+import { useEffect } from "react";
 
-export const Route = createFileRoute('/(auth)/logout')({
-  component: RouteComponent,
-})
+export const Route = createFileRoute("/(auth)/logout")({
+  component: Logout,
+});
 
-function RouteComponent() {
-  return <div>Hello "/(auth)/logout"!</div>
+function Logout() {
+  const redirect = useNavigate();
+  
+  useEffect(() => {
+    localStorage.removeItem("auth.user_id");
+    localStorage.removeItem("auth.displayName");
+    localStorage.removeItem("auth.token");
+    redirect({ to: "/login" });
+  }, [redirect]);
+
+  return <></>;
 }
